Expanded ResourceService

+++ crisischeckin/Services/DataService.cs
@@ -2,6 +2,8 @@
 using Services.Exceptions;
 using Services.Interfaces;
 using System;
+using System.Collections.Generic;
+using System.Data.Entity;
 using System.Linq;
 using System.Threading.Tasks;
 using Models.Migrations;
@@ -139,6 +141,8 @@ namespace Services
             return result;
         }
 
+
+
         public void AddContact(Contact newContact)
         {
             context.Contacts.Add(newContact);
@@ -151,18 +155,44 @@ namespace Services
             await context.SaveChangesAsync();
         }
 
+        public async Task<Resource> FindResourceByIdAsync(int? id)
+        {
+            var resource = await context.Resources.Where(r => r.ResourceId == id)
+                                    .Include(r => r.Person)
+                                    .Include(r => r.ResourceType)
+                                    .FirstOrDefaultAsync();
+
+            return resource;
+        }
+
+        public async Task RemoveResourceByIdAsync(int id)
+        {
+            var resource = await context.Resources.FindAsync(id);
+
+            if (resource == null)
+            {
+                throw new ResourceNotFoundException();
+            }
+
+            context.Resources.Remove(resource);
+            await context.SaveChangesAsync();
+        }
+
+        public async Task<IEnumerable<Resource>> GetAllResourcesAsync()
+        {
+            return await context.Resources.Include(r => r.Disaster)
+                                               .Include(r => r.Person)
+                                               .Include(r => r.ResourceType)
+                                               .ToListAsync();
+        }
+
         public void AddCommitment(Commitment newCommitment)
         {
             context.Commitments.Add(newCommitment);
             context.SaveChanges();
         }
 
-        public async Task RemoveResourceByIdAsync(int id)
-        {
-            var resource = await context.Resources.FindAsync(id);
-            context.Resources.Remove(resource);
-            await context.SaveChangesAsync();
-        }
+
 
         public void RemoveCommitmentById(int id)
         {
